Add solution for problem `A - Integer Moves` Link: https://codeforces.com/contest/1657/problem/A Tags: brute force, math, *800 Ref: https://codeforces.com/contest/1657/submission/150785552

+++ codeforces/1657/A.cpp
@@ -8,7 +8,7 @@ int main()
    cin >> t;
    while (t--)
    {
-      long long x, y;
+      int x, y;
       cin >> x >> y;
       if (x == 0 && y == 0)
       {
@@ -16,7 +16,7 @@ int main()
       }
       else
       {
-         long long a = pow(((x * x) + (y * y)), 0.5);
+         int a = pow(((x * x) + (y * y)), 0.5);
          if ((a *a) == ((x * x) + (y * y)))
          {
             cout << 1 << endl;
